organize block explorer comp
