+++ src/components/SpBlockExplorer.vue
@@ -29,9 +29,9 @@
 import { mapGetters, mapMutations, mapActions } from 'vuex'
 import { formatter as blockFormatter } from '../helpers/block'
 
-import FullWidthContainer from './layouts/FullWidthContainer'
-import BlockDetailSheet from './blocks/BlockDetailSheet'
-import BlockChain from './blocks/BlockChain'
+import FullWidthContainer from './SpBlockExplorer/FullWidthContainer'
+import BlockDetailSheet from './SpBlockExplorer/BlockDetailSheet'
+import BlockChain from './SpBlockExplorer/BlockChain'
 import IconBox from '../assets/icons/Box'
 
 export default {
